Move characters based solely on their direction now
We also removed backup_direction in the process and introduced a new active property. If the character is not active (i.e, immediately after firing), they cannot move. Also a pointing direction property was added so NO_DIRECTION is not used for indicating invalid direction anymore, just not currently moving instead.