Rewrite networking code to resolve race conditions and isolate state
Race conditions are resolved by using a queue for messages going to our server/client thread and a queue for messages going to the main thread. Locks are used for appending and popping elements from the queues.

We also isolated sendto() to be only inside our net code threads, and we now wait/join for a thread to finish when exiting a game or exiting the app.