Add packet numbering and packet resending for critical messages
For non-critical messages (movement messages we send all the time), we discard packets that were meant to be sent earlier.

For critical messages, we ensure that we receive them in the order they were sent. We resend messages until we receive a corresponding ack message.

The quit message is not critical nor has a packet number attached.