Adjust aspect ratio based on drawable size
We still try to stretch background textures so that they fill the entire screen, but now objects in the game shouldn't look as weird (eg, it feeling like moving vertically is slower than horizontally if played on a wide screen).

On iOS on some phone devices, this should also give more room for using touch controls.